Major performance improvement
The change should be a no-op in terms of what the function that I
changed returns. The obj argument of this recursive function is not
actually used anywhere --it only appears in the recursive call and not
at the base cases.

`str(obj)` is a very expensive operation that we end up
needlessly going through for O(rows) operations while initialising a
DataGrid.

This change significantly improves performance for me: Initialising a
DataGrid for a DataFrame of 30k rows goes from 2.4s to 0.4s.

Signed-off-by: Vasilis Themelis <vdthemelis@gmail.com>